refactor: quantization for linear layers
